# Use shorter values for indices in delaunator-cpp