I had a quick stab at #3935 just to see how it would work... It is a pure search & replace of `std::size_t` to a new typedef `delaunator::index_t` (which is now defined as `uint32_t`). Haven't checked the code thoroughly.

Maybe there should be some errors thrown in case someone tried using more than 4 billion points?

Some quick results from a test with ~10M point LAS dataset:

|  |  Mem use before (MB) | Mem use after (MB) | Mem `after-before` (MB) | Time (s) |
|-|-|-|-|-|
| size_t | 928.3 | 2409.3 | 1481.0 | 11.0 |
| uint32_t | 928.3 | 1865.9 | 937.6 | 10.3 |

Memory usage from massif (before delaunator starts and after it initializes all vectors), timing being average from 5 runs, runing on my laptop with 64-bit Intel CPU (linux).

Testing pipeline:
```json
[
    "test.las",
    {
        "type": "filters.delaunay"
    }
]
```


Overall, with uint32_t the memory consumption of delaunator was ~37% less compared to size_t and there's a modest speed improvement (~6%).