@abellgithub 2023-01-12
People do have point clouds that exceed the size supported by the proposed change. Perhaps we should just check before use and report an error?