@wonder-sk 2023-01-12
> Perhaps we should just check before use and report an error?

Yeah - or possibly we could have templated delaunator with variants for both uint32_t and size_t, and PDAL would use the right variant? It is not that much code I guess so it should not be a big deal to have it compiled twice in the binaries...